requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
constantly==15.1.0
Django==3.1.3
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
lxml==4.6.1
PyHamcrest==2.0.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
zope.interface==5.2.0
